# swagger/swagger.yml
info:
  title: Comentario
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
swagger: "2.0"
basePath: /api
securityDefinitions:
  # Cookie authentication for commenters
  commenterTokenHeader:
    type: apiKey
    in: header
    name: X-Commenter-Token

  # Cookie authentication for owners. Uses the apiKey type for the lack of a proper cookie-based authentication in
  # Swagger 2
  ownerCookie:
    type: apiKey
    in: header
    name: Cookie
definitions:
...
  commenter:
    type: object
    properties:
      commenterHex:
        $ref: "#/definitions/hexId"
      email:
        type: string
        format: email
      name:
        type: string
      link:
        type: string
      photo:
        type: string
      provider:
        type: string
      joinDate:
        type: string
        format: date-time
      isModerator:
        type: boolean
        x-omitempty: false
...
  email:
    type: object
    properties:
      email:
        type: string
        format: email
      unsubscribeSecretHex:
        $ref: "#/definitions/hexId"
      lastEmailNotificationDate:
        type: string
        format: date-time
      sendReplyNotifications:
        type: boolean
        x-omitempty: false
      sendModeratorNotifications:
        type: boolean
        x-omitempty: false
...
  hexId:
    description: ID consisting of 64 hex digits
    type: string
    minLength: 64
    maxLength: 64
    pattern: '[0-9a-f]{64}'
...
paths:
...
  /commenter/login:
    post:
      operationId: CommenterLogin
      summary: Sign in as specified commenter
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
                format: email
              password:
                type: string
                minLength: 1
      responses:
        200:
          description: Logged in successfully
          schema:
            type: object
            properties:
              commenterToken:
                $ref: "#/definitions/hexId"
              commenter:
                $ref: "#/definitions/commenter"
              email:
                $ref: "#/definitions/email"
  /commenter/logout:
    post:
      operationId: CommenterLogout
      summary: Log the currently signed-in commenter out
      security:
        - commenterTokenHeader: []
      responses:
        204:
          description: Logged out successfully
  /commenter/new:
    post:
      operationId: CommenterNew
      summary: Sign up as a new commenter
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - name
              - email
              - password
            properties:
              name:
                type: string
                minLength: 1
                maxLength: 63
              website:
                type: string
                minLength: 1
                maxLength: 2083
              email:
                type: string
                format: email
              password:
                type: string
                minLength: 1
                maxLength: 63
      responses:
        204:
          description: Commenter signed up successfully
  /commenter/photo:
    get:
      operationId: CommenterPhoto
      summary: Get an avatar for given commenter in JPEG format
      produces:
        - image/jpeg
      parameters:
        - name: commenterHex
          in: query
          type: string
          required: true
          minLength: 64
          maxLength: 64
          pattern: '[0-9a-f]{64}'
      responses:
        200:
          description: Success or failure response (see 'success' property)
          schema:
            type: file
  /commenter/self:
    post:
      operationId: CommenterSelf
      summary: Request commenter's sign-in status and, if the commenter's signed in, their details
      responses:
        204:
          description: Commenter is not signed in
        200:
          description: Commenter is signed in
          schema:
            type: object
            properties:
              commenter:
                $ref: "#/definitions/commenter"
              email:
                $ref: "#/definitions/email"
...
  /commenter/update:
    post:
      operationId: CommenterUpdate
      summary: Update commenter details
      security:
        - commenterTokenHeader: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - name
              - email
            properties:
              name:
                type: string
                minLength: 1
                maxLength: 63
              email:
                type: string
                format: email
              link:
                type: string
                minLength: 1
                maxLength: 2083
              photo:
                type: string
                minLength: 1
                maxLength: 2083
      responses:
        204:
          description: Commenter details haven been updated

  #---------------------------------------------------------------------------------------------------------------------
  # Domains
  #---------------------------------------------------------------------------------------------------------------------
...